update commit - readme.md

+++ driver/software/RF212b/phy/at86rf212b/src/phy_tx.c
@@ -294,7 +294,7 @@ void handle_tx_end_irq(bool underrun_occured)
 				trx_delay_micros(1);
 				TRX_SLP_TR_LOW();
 				uint8_t context = 0;
-    			trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, &context, 54000, SYS_TIME_SINGLE);
+    			trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, (uintptr_t)&context, 54000, SYS_TIME_SINGLE);
     			if(trxEIC_waitTimer == SYS_TIME_HANDLE_INVALID)
     			{
             		return;
