+++ driver/software/RF233/phy/at86rf233/src/phy_tx.c
@@ -233,8 +233,8 @@ void send_frame(PHY_CSMAMode_t csmaMode, bool txRetries)
 	trx_delay_micros(1);
     TRX_SLP_TR_LOW();
 
-    uintptr_t context = 0;
-    trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, &context, 54000, SYS_TIME_SINGLE);
+    uint8_t context = 0;
+    trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, (uintptr_t)&context, 54000, SYS_TIME_SINGLE);
     if(trxEIC_waitTimer == SYS_TIME_HANDLE_INVALID)
     {
             return;
@@ -278,8 +278,8 @@ void handle_tx_end_irq(bool underrun_occured)
 				TRX_SLP_TR_HIGH();
 				trx_delay_micros(1);
 				TRX_SLP_TR_LOW();
-				uintptr_t context = 0;
-    			trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, &context, 54000, SYS_TIME_SINGLE);
+				uint8_t context = 0;
+    			trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, (uintptr_t)&context, 54000, SYS_TIME_SINGLE);
     			if(trxEIC_waitTimer == SYS_TIME_HANDLE_INVALID)
     			{
             		return;
